Remove unnecessary part from `E1-572G.aml`
Remove a part of the E1-572G.aml that disables the EHCI Controller. Initially, this was added in an attempt to fix sleep issues, but it turns out it wasn't the culprit at all.

If you're curious about what the issue was, it was some methods in (1) DSDT and the (2) RTC. This was resolved by SSDT-GPRW, and RTCMemoryFixup + RTC blacklist.